Removed redundancy. Used reusable ui-components
