+++ src/screens/home/favorite-currency-list/favorite-currency-list-screen.js
@@ -2,16 +2,27 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FlatList, StatusBar } from 'react-native';
 
-import { FavoriteListItem } from './favorite-list-item';
-import { Separator } from '../../../ui-components/list-item';
-import { StyledView } from './styled';
+import { Separator, ListItem } from '../../../ui-components/list-item';
+import { StyledView, StyledTouchableHighlight, StyledIcon } from './styled';
+import { ICON_SIZE, BLUE, UNDERLAY_COLOR } from '../../../theme';
 
 export const FavoriteCurrencyList = ({ favoriteCurrencyList, onUnMarkCurrencyFromFavorites }) => (
   <StyledView>
     <StatusBar translucent={false} barStyle='light-content' />
     <FlatList
       data={favoriteCurrencyList}
-      renderItem={({ item }) => <FavoriteListItem item={item} onPress={() => onUnMarkCurrencyFromFavorites(item)} />}
+      renderItem={({ item }) => (
+        <ListItem
+          text={`1 ${item.baseCurrency} = ${item.conversionRate} ${item.quoteCurrency} as of ${item.lastConvertedDate}`}
+          customIcon={
+            <StyledTouchableHighlight
+              onPress={() => onUnMarkCurrencyFromFavorites(item)}
+              underlayColor={UNDERLAY_COLOR}>
+              <StyledIcon name='heart-o' size={ICON_SIZE + 5} color={BLUE} />
+            </StyledTouchableHighlight>
+          }
+        />
+      )}
       keyExtractor={item => item.conversionRate.toString()}
       ItemSeparatorComponent={Separator}
     />
