--- src/screens/home/favorite-currency-list/favorite-list-item/index.js
@@ -1,1 +1,0 @@
-export * from './favorite-list-item';
